server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
google-genai==1.49.0

